requirements.txt
numpy<2.0.0
torch==2.3.0
opencv-python==4.9.0.80
onnxruntime==1.12.0
nuscenes-devkit==1.1.11
loguru==0.7.2